Show playa link instead of stats when not on the playa
