+++ src/components/molecules/NavBar/NavBar.scss
@@ -4,6 +4,7 @@
   background: $black;
   display: flex;
   justify-content: space-between;
+  align-items: center;
 
   .log-in-button {
     cursor: pointer;
